cleanup: remove dead parser and helper code
